requirements.txt
Flask==2.3.2
flask_socketio==5.5.1
eventlet==0.33.3
gunicorn==21.2.0
python-shogi==1.1.1